requirements.txt
tweepy>=4.14.0
feedparser>=6.0.11
mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
